requirements.txt
discord.py==2.5.2
python-dotenv==1.1.1
PyYAML==6.0.2
rcon==2.4.9
